API: fix update ticket with change of status validation. APP: refresh ticket status counters

+++ app/src/components/Board/Board.js
@@ -93,31 +93,40 @@ class Board extends React.Component {
         }
     }
     async update(ticket) {
+        const oldTicket = this.state.tickets.filter(t => t._id === ticket._id)[0];
         await this.props.modifyTicket(ticket);
-        this.setState((prevState, props) => {
-            //console.log(props)
-            return {
-                ...prevState,
-                tickets: props.tickets,
-                showForm: false,
-                //ticketToEdit: {}
-            }
-        });
-    }
-    async deleteTicket(ticket) {
-        await this.props.deleteTicketById(ticket._id)
-        this.setState((prevState, props) => {
-            return {
-                ...prevState,
-                tickets: props.tickets,
-                ticketToEdit: {},
-                showForm: false,
-                ticketCounterStatus: {
-                    ...prevState.ticketCounterStatus,
-                    [ticket.status.name]: prevState.ticketCounterStatus[ticket.status.name] - 1
+        if (!this.props.error) {
+            this.setState((prevState, props) => {
+                return {
+                    ...prevState,
+                    tickets: props.tickets,
+                    showForm: false,
+                    ticketCounterStatus: {
+                        ...prevState.ticketCounterStatus,
+                        [oldTicket.status.name.toLowerCase()]: prevState.ticketCounterStatus[oldTicket.status.name.toLowerCase()] - 1,
+                        [ticket.status.name.toLowerCase()]: prevState.ticketCounterStatus[ticket.status.name.toLowerCase()] + 1
+                    }
                 }
-            }
-        });
+            });
+        }
+    }
+    async deleteTicket(e, ticket) {
+        e.preventDefault();
+        if (window.confirm(`¿Está seguro que desea borrar ${ticket.number} ${ticket.title}?`)) {
+            await this.props.deleteTicketById(ticket._id)
+            this.setState((prevState, props) => {
+                return {
+                    ...prevState,
+                    tickets: props.tickets,
+                    ticketToEdit: {},
+                    showForm: false,
+                    ticketCounterStatus: {
+                        ...prevState.ticketCounterStatus,
+                        [ticket.status.name.toLowerCase()]: prevState.ticketCounterStatus[ticket.status.name.toLowerCase()] - 1
+                    }
+                }
+            });
+        }
     }
     render() {
         return (
@@ -126,6 +135,7 @@ class Board extends React.Component {
                     <h1 className="title-page"><Icon>business_center</Icon> Proyecto: <a href={`/projects/${this.props.projectId}`}>{this.props.project.name}</a></h1>
                     <Col s={12} className="grid-center">
                         { this.state.isLoading && (<Preloader />)}
+                        { this.props.error && <p className="error-msg"><Icon left>error</Icon> Error al actualizar el ticket. Revisar el estado asignado</p> }
                     </Col>
                     <Col s={12} m={12} l={12}>
                         <TicketForm show={this.state.showForm}
@@ -133,8 +143,7 @@ class Board extends React.Component {
                                     statuses={this.props.statuses}
                                     ticket={this.state.ticketToEdit}
                                     handleCreate={this.addTicket}
-                                    handleUpdate={this.update}
-                                    handleDelete={this.deleteTicket} />
+                                    handleUpdate={this.update} />
                     </Col>
                     {this.state.statuses && (this.state.statuses.map(status => 
                     <Col key={status._id} s={12} m={4} l={2} className="container-app column-stattus">
@@ -143,13 +152,14 @@ class Board extends React.Component {
                             <span className="counter">{this.state.ticketCounterStatus[status.name]}</span>   
                         </div>
                         <hr />
-                        {this.state.tickets.length > 0 && this.state.tickets.map(ticket => {
+                        {this.state.tickets && this.state.tickets.map(ticket => {
                             return ticket.status.name.toLowerCase() === status.name.toLowerCase() && (
                                 <Card key={ticket._id}
                                     title={ticket.title.length >= 40 ? ticket.title.substring(0, 40) + " ..." : ticket.title}
                                     actions={[
                                         <a key="1" href="/" onClick={e => this.openForm(e, ticket)}>Editar</a>,
-                                        <span key="2" className="number"># {ticket.number}</span>
+                                        <a key="2" href="/" onClick={e => this.deleteTicket(e, ticket)}>Borrar</a>,
+                                        <span key="3" className="number"># {ticket.number}</span>
                                     ]}>
                                     <hr className={"divider-app"} />
                                     <p><i>Estado: {ticket.status.name.toUpperCase()}</i></p><br />
